#2871 fix all test cases, especially nest and unnest cases
